requirements.txt
psycopg2-binary>=2.9.0
# Add other dependencies here if needed (currently the project uses mostly standard libraries)
